+++ test/models/solid_agents/run_test.rb
@@ -10,7 +10,7 @@ module SolidAgents
       assert_equal "queued", run.status
       assert_equal "received", run.stage
       assert_equal "alex", run.stage_owner
-      assert_equal "pi", run.runtime
+      assert_equal "ruby_llm", run.runtime
       assert_equal "bin/rails test", run.test_command
     end